Adding Boot-Time Diagnostics
So given that within the internet, nothing is guaranteed, it's important to have safeguards everywhere, and explicit error messages.
At this point, here's what was added:

- When the module is required for the first time within the game, a couple checks will be put in place:
- - Whether HttpService is enabled;
- - Whether the key is valid;
- - Whether the service is stable (API Latency and sample status codes).

This way, it will be easier to identify and diagnose on-set up issues.
In a future development we should also do this within sending requests themselves to avoid malformed data to reach processing structures.

I also modified the auth module to clearly split the KEY and TOKEN fields.
- KEY identifies the Trello account/developer;
- TOKEN will allow for more advanced features. Seemingly it's not needed if all you want is to read from a public board.

The default.project.json file is a Rojo project file.

Signed-off-by: David Duque <david.f.s.duque@tecnico.ulisboa.pt>